# Conflicts:
#	backend-nest/src/api/persist/persist.service.ts
#	frontend-react/src/app/login/page.tsx

+++ backend-nest/src/api/users/users.service.ts
@@ -1,4 +1,9 @@
-import { Injectable } from '@nestjs/common';
+import {
+  ConflictException,
+  Injectable,
+  InternalServerErrorException,
+  NotFoundException,
+} from '@nestjs/common';
 import { PrismaService } from '../../prisma/prisma.service';
 import { Prisma, user } from '@prisma/client';
 import { CreateUserDto } from './dto/create-user.dto';
@@ -11,15 +16,25 @@ export class UsersService {
   async findOne(
     userWhereUniqueInput: Prisma.userWhereUniqueInput,
   ): Promise<user | null> {
-    return this.prisma.user.findUnique({
-      where: userWhereUniqueInput,
-    });
+    try {
+      return await this.prisma.user.findUnique({
+        where: userWhereUniqueInput,
+      });
+    } catch (error) {
+      console.error(error);
+      throw new InternalServerErrorException('Error finding user');
+    }
   }
 
   async findOneByEmail(email: string): Promise<user | null> {
-    return this.prisma.user.findFirst({
-      where: { email },
-    });
+    try {
+      return await this.prisma.user.findFirst({
+        where: { email },
+      });
+    } catch (error) {
+      console.error(error);
+      throw new InternalServerErrorException('Error finding user by email');
+    }
   }
 
   async findAll(params: {
@@ -30,19 +45,36 @@ export class UsersService {
     orderBy?: Prisma.userOrderByWithRelationInput;
   }): Promise<user[]> {
     const { skip, take, cursor, where, orderBy } = params;
-    return this.prisma.user.findMany({
-      skip,
-      take,
-      cursor,
-      where,
-      orderBy,
-    });
+    try {
+      return await this.prisma.user.findMany({
+        skip,
+        take,
+        cursor,
+        where,
+        orderBy,
+      });
+    } catch (error) {
+      console.error(error);
+      throw new InternalServerErrorException('Error finding users');
+    }
   }
 
   async create(data: CreateUserDto): Promise<user> {
-    return this.prisma.user.create({
-      data,
-    });
+    const existingUser = await this.findOneByEmail(data.email);
+    if (existingUser) {
+      const errorMessage = 'An account with this email already exists';
+      console.error(errorMessage);
+      throw new ConflictException(errorMessage);
+    }
+
+    try {
+      return await this.prisma.user.create({
+        data,
+      });
+    } catch (error) {
+      console.error('Error when creating user :', error);
+      throw new InternalServerErrorException('Error when creating user :');
+    }
   }
 
   async update(params: {
@@ -50,22 +82,66 @@ export class UsersService {
     data: UpdateUserDto;
   }): Promise<user> {
     const { where, data } = params;
-    return this.prisma.user.update({
-      data,
-      where,
-    });
+    const existingUser = await this.findOne(where);
+    if (!existingUser) {
+      throw new NotFoundException('User not found');
+    }
+
+    const userWithUpdatedEmail = await this.findOneByEmail(data.email);
+    if (userWithUpdatedEmail && userWithUpdatedEmail.id !== existingUser.id) {
+      throw new ConflictException('Email already in use');
+    }
+
+    try {
+      return await this.prisma.user.update({
+        data,
+        where,
+      });
+    } catch (error) {
+      console.error("Erreur lors de la mise à jour de l'utilisateur :", error);
+      throw new InternalServerErrorException(
+        "Erreur lors de la mise à jour de l'utilisateur.",
+      );
+    }
   }
 
   async delete(where: Prisma.userWhereUniqueInput): Promise<user> {
-    return this.prisma.user.delete({
-      where,
-    });
+    const existingUser = await this.findOne(where);
+    if (!existingUser) {
+      throw new NotFoundException('User not found');
+    }
+
+    try {
+      return await this.prisma.user.delete({
+        where,
+      });
+    } catch (error) {
+      console.error("Erreur lors de la suppression de l'utilisateur :", error);
+      throw new InternalServerErrorException(
+        "Erreur lors de la suppression de l'utilisateur.",
+      );
+    }
   }
 
   async updateClassGroup(userId: number, classGroupId: number): Promise<user> {
-    return this.prisma.user.update({
-      where: { id: userId },
-      data: { classGroupId },
-    });
+    const existingUser = await this.findOne({ id: userId });
+    if (!existingUser) {
+      throw new NotFoundException('User not found');
+    }
+
+    try {
+      return await this.prisma.user.update({
+        where: { id: userId },
+        data: { classGroupId },
+      });
+    } catch (error) {
+      console.error(
+        "Erreur lors de la mise à jour du groupe de classe de l'utilisateur :",
+        error,
+      );
+      throw new InternalServerErrorException(
+        "Erreur lors de la mise à jour du groupe de classe de l'utilisateur.",
+      );
+    }
   }
 }
